// apps/git-client/src/generated/TerminalEvent.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type TerminalEvent = { "kind": "output", sequence: number, data: Array<number>, } | { "kind": "exited", exitCode: number, signal: string | null, } | { "kind": "failed", message: string, };